Handle events in app object
Prior to this commit we were using events.js to handle application
events, but app has events to so we need to change our application
to remove the events and start using app to handle events.

We accomplish this by exporting the app object and making it
available in app.js so other modules can use it.